# docs/swagger.yaml
basePath: /api
definitions:
  lab_internal_app_DTO.AddSolarPanel:
    properties:
      depth:
        type: integer
      description:
        type: string
      efficiency:
        type: string
      height:
        type: integer
      power:
        type: number
      title:
        type: string
      type:
        type: string
      width:
        type: integer
    required:
    - depth
    - efficiency
    - height
    - power
    - title
    - type
    - width
    type: object
  lab_internal_app_DTO.ChangeSolarPanel:
    properties:
      depth:
        type: integer
      description:
        type: string
      efficiency:
        type: string
      height:
        type: integer
      power:
        type: number
      title:
        type: string
      type:
        type: string
      width:
        type: integer
    type: object
...
  lab_internal_app_ds.SolarPanel:
    properties:
      depth:
        type: integer
      description:
        type: string
      efficiency:
        type: string
      height:
        type: integer
      id:
        type: integer
      image:
        type: string
      isDelete:
        type: boolean
      power:
        type: number
      title:
        type: string
      type:
        type: string
      width:
        type: integer
    type: object
...
host: localhost:8001
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API для управления солнечными панелями и расчета их мощности
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Solar Panels Power Calculator API
  version: "1.0"
paths:
...
  /panels:
    get:
      consumes:
      - application/json
      description: Возвращает список всех солнечных панелей с возможностью фильтрации
        по мощности
      parameters:
      - description: Минимальная мощность для фильтрации
        in: query
        name: start_value
        type: number
      - description: Максимальная мощность для фильтрации
        in: query
        name: end_value
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: Список солнечных панелей
          schema:
            items:
              $ref: '#/definitions/lab_internal_app_ds.SolarPanel'
            type: array
        "400":
          description: Некорректные параметры фильтрации
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Панели не найдены
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить список солнечных панелей
      tags:
      - SolarPanels
    post:
      consumes:
      - application/json
      description: Создает новую солнечную панель. Требуется роль модератора.
      parameters:
      - description: Данные новой панели
        in: body
        name: solarPanel
        required: true
        schema:
          $ref: '#/definitions/lab_internal_app_DTO.AddSolarPanel'
      produces:
      - application/json
      responses:
        "201":
          description: Созданная панель
          schema:
            $ref: '#/definitions/lab_internal_app_ds.SolarPanel'
        "400":
          description: Некорректные данные
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Недостаточно прав
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Добавить новую солнечную панель (Модератор)
      tags:
      - SolarPanels
  /panels/{id}:
    delete:
      consumes:
      - application/json
      description: Помечает солнечную панель как удаленную. Требуется роль модератора.
      parameters:
      - description: ID солнечной панели
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Панель успешно удалена
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Недостаточно прав
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Панель не найдена
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Удалить солнечную панель (Модератор)
      tags:
      - SolarPanels
    get:
      consumes:
      - application/json
      description: Возвращает детальную информацию о конкретной солнечной панели
      parameters:
      - description: ID солнечной панели
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Данные солнечной панели
          schema:
            $ref: '#/definitions/lab_internal_app_ds.SolarPanel'
        "404":
          description: Панель не найдена
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить солнечную панель по ID
      tags:
      - SolarPanels
    post:
      consumes:
      - application/json
      description: Добавляет солнечную панель в текущую заявку пользователя (корзину)
      parameters:
      - description: ID солнечной панели
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Панель добавлена в заявку
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Панель уже в заявке или удалена
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Панель не найдена
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Добавить панель в корзину
      tags:
      - SolarPanels
    put:
      consumes:
      - application/json
      description: Обновляет данные существующей солнечной панели. Требуется роль
        модератора.
      parameters:
      - description: ID солнечной панели
        in: path
        name: id
        required: true
        type: integer
      - description: Обновленные данные панели
        in: body
        name: solarPanel
        required: true
        schema:
          $ref: '#/definitions/lab_internal_app_DTO.ChangeSolarPanel'
      produces:
      - application/json
      responses:
        "200":
          description: Обновленная панель
          schema:
            $ref: '#/definitions/lab_internal_app_ds.SolarPanel'
        "400":
          description: Некорректные данные
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Недостаточно прав
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Панель не найдена
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Изменить солнечную панель (Модератор)
      tags:
      - SolarPanels
  /panels/{id}/image:
    post:
      consumes:
      - multipart/form-data
      description: Загружает изображение для солнечной панели. Требуется роль модератора.
      parameters:
      - description: ID солнечной панели
        in: path
        name: id
        required: true
        type: integer
      - description: Файл изображения
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Панель с обновленным изображением
          schema:
            $ref: '#/definitions/lab_internal_app_ds.SolarPanel'
        "400":
          description: Файл не найден
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Недостаточно прав
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Панель не найдена
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Загрузить изображение панели (Модератор)
      tags:
      - SolarPanels
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"